hud_chas

This function queries the CHAS API provided by US Department of Housing and Urban Development

hud_chas(
  type,
  stateId = NULL,
  entityId = NULL,
  year = c("2014-2018"),
  key = Sys.getenv("HUD_KEY")
)

Arguments

type

Queries the data based off: 1 - Nation 2 - State 3 - County 4 - MCD 5 - Place

stateId

For types 2,3,4,5, you must provide a stateId. For 3,4,5

entityId

For types 3,4,5, you must provide a fips code

year

Gets the year that this data was recorded. Defaults to 2014-2018. There are specific year ranges that are only accepted. 2014-2018 2013-2017 2012-2016 2011-2015 2010-2014 2009-2013 2008-2012 2007-2011 2006-2010

key

The API key for this user. You must go to HUD and sign up for an account and request for an API key.

Value

This function returns a dataframe containing CHAS data for a particular state. For more details about these measurements, go to https://www.huduser.gov/portal/dataset/chas-api.html

Examples

if (FALSE) {
library(hudr)

Sys.setenv("HUD_KEY" = "q3r2rjimd129fj121jid")

hud_chas(1)

hud_chas('2', stateId = '56')

hud_chas('3','51','199')

hud_chas('4', '51', 94087)

hud_chas('5', '51', 48996)
}